+++ public/controls/pictures.js
@@ -1,6 +1,6 @@
 import { io } from "https://cdn.socket.io/4.4.1/socket.io.esm.min.js";
 const socket = io();
-
+import tdate from "./tdate.js";
 export default class BudgetTracker {
   constructor(querySelectorString) {
     this.root = document.querySelector(querySelectorString);
@@ -22,14 +22,15 @@ export default class BudgetTracker {
   static entryHtml() {
     return `
             <tr>
-                <td>
-                <button type="button"  class="btn-rounded" id="btnDel" style="position: absolute;z-index: 2;right:10px;border-radius: 50%;"><i id="btnData" class="fa fa-trash" aria-hidden="true"></i></button>
-                  <div id="picture" ></div>
+                <td style="height:300px">
+                    <div id="title"></div>
+                    <button type="button"  class="btn-rounded" id="btnDel" style="position: absolute;z-index: 2;right:10px;border-radius: 50%;"><i id="btnData"  class="fa fa-trash" aria-hidden="true"></i></button>
+                    <button type="button"  class="btn-rounded" id="btnDownload" style="position: absolute;z-index: 2;right:50px;border-radius: 50%;"><i id="btnDataDownload"  class="fa fa-download" aria-hidden="true"></i></button>
+                    <div id="picture"></div>
                 </td>
             </tr>
         `;
   }
-
   load(entries) {
     for (const entry of entries) {
       this.addEntry(entry);
@@ -41,6 +42,7 @@ export default class BudgetTracker {
       .querySelector(".entries")
       .insertAdjacentHTML("afterbegin", BudgetTracker.entryHtml());
     let row = this.root.querySelector(".entries tr:first-of-type");
+    let title = this.root.querySelector("#title");
     let type = entry.type.split("/");
 
     if (type[0] === "image") {
@@ -54,6 +56,7 @@ export default class BudgetTracker {
         </div>
         </div>
             ` || "";
+      title.innerHTML = ` ${entry.username}[news] ${tdate(new Date())}`;
 
       let src = `uploads/${entry.filename}`;
       loadImage(src).then((images) => {
@@ -84,12 +87,29 @@ export default class BudgetTracker {
             ` || "";
     }
     row.querySelector("#btnData").dataset.pic = entry.filename;
+    row.querySelector("#btnDel").dataset.pic = entry.filename;
+    row.querySelector("#btnDataDownload").dataset.pic = entry.filename;
     row.querySelector("#btnDel").addEventListener("click", (e) => {
-      this.onDeleteEntryBtnClick(e);
+      let text = "(ต้องการลบข้อมุลหรือไม่!";
+      if (confirm(text) == true) this.onDeleteEntryBtnClick(e);
+    });
+    row.querySelector("#btnDownload").addEventListener("click", (e) => {
+      console.log("bonclick");
+      this.onDownloadEntryBtnClick(e);
     });
   }
 
   save(e) {}
+
+  onDownloadEntryBtnClick(e) {
+    let pname = e.target.dataset.pic;
+    console.log(pname);
+    if (!pname) return;
+    let a = document.createElement("a");
+    a.href = "/uploads/" + pname;
+    a.download = pname;
+    a.click();
+  }
 
   onDeleteEntryBtnClick(e) {
     if (!e.target.dataset.pic) return;
